Extracting manoeuvre-competency into comon components
