+++ src/pages/test-report/components/manoeuvre-competency/manoeuvre-competency.scss
